Failed if gemrc was already toast

+++ bootstrap.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -e
 pacman -S --needed git ruby base-devel --noconfirm
-rm /etc/gemrc
+[ -e /etc/gemrc ] && rm /etc/gemrc
 gem install chef --no-ri --no-rdoc
 
 cd /tmp/chef-bib
